+++ mayan/apps/metadata/forms.py
@@ -4,6 +4,8 @@ from django.forms.formsets import formset_factory
 from django.utils.text import format_lazy
 from django.utils.translation import ugettext_lazy as _
 
+from mayan.apps.templating.fields import TemplateField
+
 from .classes import MetadataLookup
 from .models import DocumentTypeMetadataType, MetadataType
 
@@ -163,11 +165,16 @@ DocumentMetadataRemoveFormSet = formset_factory(
 class MetadataTypeForm(forms.ModelForm):
     def __init__(self, *args, **kwargs):
         super(MetadataTypeForm, self).__init__(*args, **kwargs)
-        self.fields['lookup'].help_text = format_lazy(
-            '{}{}{}',
-            self.fields['lookup'].help_text,
-            _(' Available template context variables: '),
-            MetadataLookup.get_as_help_text()
+        self.fields['default'] = TemplateField(
+            initial_help_text=self.fields['default'].help_text, required=False
+        )
+        self.fields['lookup'] = TemplateField(
+                initial_help_text=format_lazy(
+                '{}{}{}',
+                self.fields['lookup'].help_text,
+                _(' Available template context variables: '),
+                MetadataLookup.get_as_help_text()
+            ), required=False
         )
 
     class Meta: